优化上传页面总进度的交互

+++ public/static/js/files/taskuplist.js
@@ -282,7 +282,7 @@
             var label = $('#totalprogress .label');
         
             if (val != null){
-                val = parseInt(val);
+                val = parseFloat(val);
 
                 progress.progressbar({value: val});
                 progress.toggleClass('disabled', false);
